+++ tests/testhelper_test.go
@@ -1,22 +1,19 @@
-package services
+package tests
 
 import (
 	"bytes"
 	"encoding/json"
-	"fmt"
+	"github.com/alecthomas/kong"
 	"github.com/dnaeon/go-vcr/cassette"
+	"github.com/dnaeon/go-vcr/recorder"
+	"github.com/octo-cli/octo-cli/services"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
 	"testing"
-	"text/tabwriter"
-	"text/template"
-
-	"github.com/alecthomas/kong"
-	"github.com/dnaeon/go-vcr/recorder"
-	"github.com/stretchr/testify/require"
 )
 
 func init() {
@@ -42,7 +39,7 @@ func startVCR(t *testing.T, recPath string) func() {
 			r.Header.Get("Accept") == i.Headers.Get("Accept")
 	})
 	require.Nil(t, err)
-	TransportWrapper = rec
+	services.TransportWrapper = rec
 	return func() {
 		t.Helper()
 		require.Nil(t, rec.Stop())
@@ -56,7 +53,7 @@ func testCmdLine(t *testing.T, fixtureName string, cmdStruct interface{}, cmdlin
 	p, e := kong.New(cmdStruct)
 
 	require.NoError(t, e)
-	Stdout = &sout
+	services.Stdout = &sout
 
 	k, e := p.Parse(cmdline)
 	require.NoError(t, e)
@@ -69,7 +66,7 @@ func testCmdLine(t *testing.T, fixtureName string, cmdStruct interface{}, cmdlin
 }
 
 func TestCreate(t *testing.T) {
-	stdout, stderr, err := testCmdLine(t, "test_issues_create", &IssuesCmd{},
+	stdout, stderr, err := testCmdLine(t, "test_issues_create", &services.IssuesCmd{},
 		`create`,
 		`--owner=octo-cli-testorg`,
 		`--repo=test-create-issue`,
@@ -90,45 +87,4 @@ func TestCreate(t *testing.T) {
 	assert.Equal(t, "test this body", got["body"])
 	assert.Len(t, got["labels"], 2)
 	assert.EqualValues(t, 1, got["milestone"].(map[string]interface{})["number"])
-}
-
-// just keeping this here until I put it somewhere else because I don't want to forget how it works.
-func TestPlayWithOutput(t *testing.T) {
-	stdout, stderr, err := testCmdLine(t, "test_issues_create", &IssuesCmd{},
-		`create`,
-		`--owner=octo-cli-testorg`,
-		`--repo=test-create-issue`,
-		`--title=test this`,
-		`--body=test this body`,
-		`--labels=label1`,
-		`--labels=label2`,
-		`--milestone=1`,
-		`--assignees=octo-cli-testuser`,
-		`--raw-output`,
-	)
-	assert.NoError(t, err)
-	assert.Empty(t, stderr.String())
-	var got map[string]interface{}
-	err = json.Unmarshal(stdout.Bytes(), &got)
-	assert.NoError(t, err)
-
-	tp, err := template.New("").Funcs(template.FuncMap{"wider": func(a, b interface{}) string { return fmt.Sprintf("%-20s%d", a, b) }}).Parse(`
-	Number	{{.number}}
-	State	{{.state}}
-	Title	{{.title}}
-	User	{{.user.login}}
-	URL	{{.html_url}}
-	CreatedAt	{{.created_at}}
-	UpdatedAt	{{.updated_at}}
-	{{if .labels}}Labels{{range .labels}}	{{.name}}
-	{{end}}{{end}}{{if .milestone}}Milestone	{{.milestone.title}}{{end}}
-	{{if .assignees}}Assignees{{range .assignees}}	{{.login}}
-	{{end}}{{end}}
-	`)
-	assert.NoError(t, err)
-	w := tabwriter.NewWriter(os.Stdout, 8, 8, 8, ' ', 0)
-	err = tp.Execute(w, got)
-	assert.NoError(t, err)
-	err = w.Flush()
-	assert.NoError(t, err)
 }
